% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot3Dim.R
\name{animation}
\alias{animation}
\title{Create an animated GIF with the rgl device}
\usage{
animation(directory = "", file_name = "animation", axis_spin = c(0, 1, 0),
  axis_spin_rpm = 5)
}
\arguments{
\item{directory}{The directory within the working directory. For example, "MyFolder/".}

\item{file_name}{The name of the output file.}

\item{axis_spin}{a numeric vector of lenght 3 indicating around which axis should the plot spin. See \code{\link[rgl]{movie3d}}.}

\item{axis_spin_rpm}{the velocity of the spin. See \code{\link[rgl]{movie3d}}.}
}
\description{
Create an animated GIF with the rgl device
}

