% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot2Dim.R
\name{biplot_2d}
\alias{biplot_2d}
\title{2D biplot}
\usage{
biplot_2d(ordination_object, ordination_method = "PCA", groups = NULL,
  vips = NULL, show_grid = TRUE, show_axes = TRUE,
  show_group_legend = FALSE, show_vip_legend = TRUE, show_arrows = TRUE,
  show_fitAnalysis = TRUE, grid_cex = 1, grid_font = 3, grid_adj = 0.5,
  invert_coordinates = c(FALSE, FALSE), xlim = NULL, ylim = NULL,
  x_title = "", y_title = "", title_cex = 1, title_font = 2,
  title_adj = 0.5, subtitle = NULL, subtitle_position = "bottomleft",
  subtitle_cex = 1, point_type = "point", point_pch = 1, point_size = 1,
  point_label = NULL, point_label_cex = 1, point_label_font = 3,
  point_label_adj = c(0.5, 0.5), point_label_adj_override = NULL,
  group_color = "black", group_star_cex = 1, group_ellipse_cex = 1,
  group_ellipse_axes = FALSE, group_label_cex = 1,
  group_legend_title = "groups", group_legend_title_pos = c(0.5, 0.85),
  group_legend_title_cex = 1, group_legend_title_font = 3,
  group_legend_title_adj = 0.5, group_legend_box_color = "white",
  group_legend_key_pch = 15, group_legend_key_cex = 1,
  group_legend_key_lwd = 1, group_legend_key_margin = 0.15,
  group_legend_text_margin = 0.25, group_legend_text_color = "black",
  group_legend_text_cex = 1, group_legend_text_font = 1,
  group_legend_text_adj = 0, vip_color = "black", vip_pch = c(0, 1, 5, 2,
  6, 4, 3), vip_cex = c(5, 5, 5, 5, 5, 3, 3), vip_lwd = 5, vip_font = 1,
  vip_adj = c(0.5, 0.5), vip_legend_title = "VIPs",
  vip_legend_title_pos = c(0.5, 0.85), vip_legend_title_cex = 2,
  vip_legend_title_font = 3, vip_legend_title_adj = 0.5,
  vip_legend_box_color = "white", vip_legend_key_cex = 0.8,
  vip_legend_key_margin = 0.15, vip_legend_text_cex = 2,
  vip_legend_text_font = 1, vip_legend_text_adj = 0,
  vip_legend_text_margin = 0.25, arrow_color = "darkorange",
  arrow_mim_dist = 0, arrow_cex = 0.1, arrow_lwd = 2,
  arrow_label_cex = 1, arrow_label_color = "black", arrow_label_font = 1,
  arrow_label_adj = c(0.5, 0.5), arrow_label_adj_override = NULL,
  fitAnalysis_cex = 1, fitAnalysis_lwd = 3,
  fitAnalysis_screePlot_color = "white",
  fitAnalysis_stress_p_color = "darkgrey",
  fitAnalysis_stress_l_color = "black", test_text = NULL, test_cex = 1,
  test_font = 1, test_adj = 0.5, main_fig = c(0, 1, 0, 1),
  group_legend_fig = c(0.8, 0.99, 0.6, 0.9), vip_legend_fig = c(0.78, 0.99,
  0.3, 0.55), arrow_fig = c(0.69, 0.99, 0.01, 0.31),
  fitAnalysis_fig = c(0.02, 0.35, 0.06, 0.25), test_fig = c(0, 0.3, 0.8, 1),
  x_title_fig = c(0.25, 1, 0.85, 1), y_title_fig = c(0.91, 1, 0, 1),
  output_type = c("preview", "png"), directory = "",
  file_name = "2D Biplot", width = 400, height = 400, family = "sans")
}
\arguments{
\item{ordination_object}{A R object containing a
direct and named reference to default ordination
outputs (i.e. \code{ordination_object$scores} or
\code{ordination_object$points},
\code{ordination_object$loadings} and
\code{ordination_object$sdev}) available for at
least 2 dimensions. Alternatively, a data frame
or matrix with at least two columns is accepted
provided that \code{ordination_method = NULL}),
which creates a two-dimensional scatter plot.}

\item{ordination_method}{Character, the ordination
method that was used to generate the ordination
object: "PCA" for Principal Components Analysis
(default), "PCoA" for Principal Coordinates Analysis,
"NMDS" for Non-metric Multidimensional Scaling, and
"LDA" for Linear Discriminant Analysis.}

\item{groups}{A factor variable containing the group
assignation of each point.}

\item{vips}{A list of logical (Boolean) vectors identifying
"Very Important Points" under different methods or criteria.}

\item{show_grid}{Logical, wheter to display the background grid.}

\item{show_axes}{Logical, wheter to display the axes.}

\item{show_group_legend}{Logical, whether to display
a legend for groups.}

\item{show_vip_legend}{Logical, whether to display a
legend for vip criteria.}

\item{show_arrows}{Logical, whether to show variable
covariance arrows.}

\item{show_fitAnalysis}{Logical, whether to display the
fit analysis plot corresponding to the ordination method
given (Scree plot or Shepard plot).}

\item{grid_cex, grid_font, grid_adj}{The scale, font type,
and text justification of the grid notation.}

\item{invert_coordinates}{Logical, vector of length two
expressing which dimensions, if any, must be inverted
before plotting (e.g. for aesthetical reasons).}

\item{xlim, ylim}{the ranges to be encompassed by the x and
y axes, if NULL they are computed
(See  \code{\link[ade4]{s.class}}).}

\item{x_title, y_title}{Character, texts to be placed in
the x and y axes.}

\item{title_cex, title_font, title_adj}{The text parameters
of the axes titles (\code{\link[graphics]{par}}).}

\item{subtitle}{Character, a subtitle to be displayed
in the bottom-left corner of the plot
(\code{csub} in \code{\link[ade4]{s.class}}).
If equals NULL, a default subtitle is displayed present the
R-Squared of the 2D fit respect to the original distances.}

\item{subtitle_position}{Character, value indicating the
position of the subtitle in the main plot ("topleft",
"topright", "bottomleft", and "bottomright"; see
\code{\link[ade4]{s.class}}).}

\item{subtitle_cex}{the font size of the
subtitle (\code{csub} in
\code{\link[ade4]{s.class}}).}

\item{point_type}{A character input accepting three
values: "point", the default \code{\link[rgl]{points3D}};
"label", displaying the content of \code{point_label};
and "point and label", placing both points and labels.}

\item{point_pch}{A number or a numerical vector given
to \code{pch} in \code{\link[graphics]{par}}.}

\item{point_size}{The size or scale given to \code{cpoint} in
\code{\link[ade4]{s.class}}.}

\item{point_label}{A character vector labelling every
observation. It's length must be equal to the number
of rows in the points/scores of the ordination object
(\code{nrow(ordination_object$points) ==
length(point_label)}).}

\item{point_label_cex, point_label_font, point_label_adj}{The text parameters of the points' labels
(\code{\link[graphics]{text}}).}

\item{point_label_adj_override}{A data frame with x,y
values to be passed to \code{adj}, overrinding the
default value given in \code{point_label_adj}.
The rows must be named exactly as points
are referred in the ordination object.}

\item{group_color}{A vector containing the colors to
be used in each group (applied to points, labels,
stars and ellipses).}

\item{group_star_cex, group_ellipse_cex, group_label_cex}{The
size or scale of the stars, ellipses, and labels
representing groups passed to \code{\link[ade4]{s.class}}
of the \code{ade4} package. Zero values render these
elements invisible.}

\item{group_ellipse_axes}{Logical, wheter to show
the ellipses axes, passed to \code{axesell} in
\code{\link[ade4]{s.class}} of the \code{ade4} package.}

\item{group_legend_title}{Character, the title of the
group legend. If equal NULL or "", no title is displayed.}

\item{group_legend_title_pos}{A numeric vector of length
two, the xy position of the title within the group legend
box.}

\item{group_legend_title_cex, group_legend_title_font, group_legend_title_adj}{The text parameters to be applied in the group legend
title (\code{\link[graphics]{par}}).}

\item{group_legend_box_color}{The background color of
the group legend box.}

\item{group_legend_key_pch, group_legend_key_cex, group_legend_key_lwd}{The type, size and line width of the keys in the
group legend.}

\item{group_legend_key_margin}{The x position of the keys
within the group legend box. Values from 0 to 1.}

\item{group_legend_text_margin}{The x position of the
text entries in the group legend. Values from 0 to 1.}

\item{group_legend_text_color}{The color or colors of
the text entries in the group legend.}

\item{group_legend_text_cex, group_legend_text_font, group_legend_text_adj}{The text parameters of the text entries in the group legend.}

\item{vip_color}{The color or colors to be used in vip markings.}

\item{vip_pch}{A character vector containing the
characters used for the vips markings under
each criterium.}

\item{vip_cex, vip_lwd, vip_font, vip_adj}{The graphical parameters of the vips markings.}

\item{vip_legend_title}{Character, the title of the vips
legend.}

\item{vip_legend_title_pos}{A numeric vector of length
two, the xy position of the title within the vips
legend box.}

\item{vip_legend_title_cex, vip_legend_title_font, vip_legend_title_adj}{The text parameters to be applied in the vips legend
title (\code{\link[graphics]{par}}).}

\item{vip_legend_box_color}{The background color of the
vips legend box.}

\item{vip_legend_key_cex}{The sizing factor of the keys
in the vips legend respect to the vips marking
in the plot.}

\item{vip_legend_key_margin}{The x position of the keys
within the vips legend box. Values from 0 to 1.}

\item{vip_legend_text_cex, vip_legend_text_font, vip_legend_text_adj}{The text parameters of the text entries in the vips legend.}

\item{vip_legend_text_margin}{The x position of the
text entries in the vips legend box. Values from 0 to 1.}

\item{arrow_color}{The color or colors to be used in
covariance arrows.}

\item{arrow_mim_dist}{The minimum distance of a arrow
from the origin of arrows (i.e. zero covariance),
in order for it to be displayed (range [0 = all arrows
are displayed,1 = no arrow is displayed]).}

\item{arrow_cex, arrow_lwd, arrow_label_cex, arrow_label_color, arrow_label_font, arrow_label_adj}{Graphical parameters of the covariance arrows and their
labels (\code{\link[graphics]{par}}).}

\item{arrow_label_adj_override}{A data frame with x,y values
to be passed to \code{adj}, overrinding the default
value given in \code{arrow_label_adj}. The rows must
be named exactly as variables are referred in the
ordination object.}

\item{fitAnalysis_cex, fitAnalysis_lwd, fitAnalysis_screePlot_color, fitAnalysis_stress_p_color, fitAnalysis_stress_l_color}{The graphical parameters of the plot for fit analysis
of the ordination method
(\code{\link[graphics]{par}},
\code{\link[vegan]{stressplot}}
of the \code{vegan} package).}

\item{test_text}{A character vector with the lines of
text presenting the results of statistical tests.}

\item{test_cex, test_font, test_adj}{The parameters of
the text with the test results
(\code{\link[graphics]{par}}).}

\item{main_fig, group_legend_fig, vip_legend_fig, fitAnalysis_fig, test_fig, x_title_fig, y_title_fig}{The
\code{fig} parameter (\code{\link[graphics]{par}})
to place in the display region of the graphics device,
respectively, the main plot, the group and vip legends,
the fit analysis plot, the tests results, and the x
and y axes titles.}

\item{output_type}{A character vector indicating the
output image types to be generated. Values accepted
are: "png", "tiff", "jpeg", "eps", and "preview"
(i.e. R graphics device).}

\item{directory}{Character, the directory within the
working directory. For example, "MyFolder/".}

\item{file_name}{Character, the name of the output file.}

\item{width, height}{Numeric, the dimensions of the output image.}

\item{family}{Character, the font family used in every text
in the plot, (\code{\link[graphics]{par}}).}
}
\description{
Generates a 2D biplot representing the default
points/scores and loadings of an ordination object, e.g. produced by
\code{\link[stats]{princomp}}).
}
\details{
This function allows customising virtually every
graphical parameter in a 2D biplot, including several extra
elements that may be useful for multivariate explorations.
Groups can be represented as stars, ellipsoids, and/or colors,
using  (\code{\link[ade4]{s.class}}), which can be tracked by a
fully-customisable legend (\code{group_legend} arguments).
Individual observations deemed exceptional (vip = Very
Important Points) can be marked with custom symbols. Whenever
is are more than one type of marking (e.g. different
methods/criteria of outlier detection), the different symbols
can be presented in a legend (\code{vip_legend} arguments).
When desired, it is possible to display a Scree plot
representing the eigenvalues (Principal Components Analysis,
Principal Coordinates Analysis) or a Shepard or stress plot
(Nonmetric Multidimensional Scaling,
e.g. \code{\link[vegan]{metaMDS}} in the vegan package) by
enabling \code{show_fit_analysis}. It is also possible to
display statistical test results by enabling \code{show.tests}
and introducing the corresponding lines in \code{test_text}.
The position of all elements (legend boxes, title and
subtitle, fit analysis plot and tests) can be customized
using the corresponding \code{fig} parameter.
}
\examples{

\dontrun{

# Use iris data
data("iris")

# get an ordination object
# ("PCA" is the default input of this function)
pca <- princomp(iris[, 1:4])

# Default plot using Species as the group factor
biplot_2d(pca, groups = iris$Species)

# ---------------------------------------------------------
# Plot groups as different point types (pch),
# make group label invisible and add a group
# legend with a a custom title.
biplot_2d(pca,
          groups = iris$Species,
          point_pch = c(1, 3, 2),
          group_cstar = 0,
          group_cellipse = 0,
          group_clabel = 0,
          show_group_legend = T,
          group_legend_title = "Species")

# ---------------------------------------------------------
# Polish covariance arrows
# Abbreviate variables names
dimnames(pca$loadings)[[1]] <- c("SL", "SW", "PL", "PW")
# Set a specific justification (adj) for each variable label
arrow_label_adj_override <- rbind(c(-0.1, 0),
                                  c(-0.1, 0.5),
                                  c(0.5, 1.3),
                                  c(0.5, 1.3))
row.names(arrow_label_adj_override) <-
   dimnames(pca$loadings)[[1]]
# Plot: arrows with different colors and
# without the background grid
biplot_2d(pca,
          groups = iris$Species,
          point_pch = c(1, 3, 2),
          group_cstar = 0,
          group_cellipse = 0,
          group_clabel = 0,
          show_group_legend = T,
          group_legend_title = "Species",
          arrow_color = c("orange",
                          "blue",
                          "red",
                          "green"),
          arrow_label_adj_override = arrow_label_adj_override,
          show_grid = FALSE)

# ---------------------------------------------------------
# Plot observations using their names
biplot_3d(pca, groups = iris$Species,
          point_type = "label", point_label = row.names(iris),
          star_label_alpha = 0,
          show_group_legend = TRUE, group_legend_title = "",
          arrow_center_pos = c(.5, 0, .5),
          arrow_body_length = 2, arrow_body_width = 2,
          show_axes = FALSE, view_zoom = 1)

# ---------------------------------------------------------
# Get arbitrary Very Important Points
irisVIP <- list(setosa = (1:nrow(iris) == 16 |
                          1:nrow(iris) == 42),
                versicolor=(1:nrow(iris) == 61),
                virginica=(1:nrow(iris) == 107 |
                           1:nrow(iris) == 118 |
                           1:nrow(iris) == 132))

# Plot observations using their names and group by Species using only color.
# Mark the VIP and add the respective legend with custom characters.
# Rotate the theta view angle to fit the arrows in the default setting.
biplot_3d(pca, groups = iris$Species,
          point_type = "label", point_label = row.names(iris),
          group_representation = NULL,
          show_group_legend = TRUE, group_legend_title = "",
          vips = irisVIP,
          vip_pch = c("X", "O", "+"), vip_cex = c(2, 2, 3),
          show_axes = FALSE, view_theta = 340)

# ---------------------------------------------------------
# Test the setosa separation
irisDist <- dist(iris[, 1:4])
setosaSeparation <- iris$Species == "setosa"
## multivariate test for the setosa separation
require(vegan)
irisTests <- NULL
irisTests$permanova <- adonis(irisDist~setosaSeparation)
irisTests$permdisp2 <- permutest(betadisper(irisDist,setosaSeparation),
                                 pairwise = TRUE)
# This function prepares a list of character vectors containing the test results
resultText <- function(tests){
  permanova_F <- as.character(round(tests$permanova$aov.tab$F.Model[1], 3))
  permanova_pvalue <- as.character(round(tests$permanova$aov.tab$"Pr(>F)"[1], 3))
  permanova_rSquared <- as.character(round(tests$permanova$aov.tab$R2[1], 3))
  permdisp2_F <- as.character(round(tests$permdisp2$tab$F[1], 3))
  permdisp2_pvalue <- as.character(round(tests$permdisp2$tab$"Pr(>F)"[1], 3))
  text <- list(c(paste("PERMANOVA:\\n   F = ", permanova_F,
                       " (p = ", permanova_pvalue, ")\\n", sep = ""),
                 c(expression(paste("   ", R^2, " =", sep = "")),
                   paste("          ", permanova_rSquared, sep = ""))),
               paste("PERMDISP2:\\n   F = ", permdisp2_F,
                     " (p = ", permdisp2_pvalue,")", sep = ""))
  return(text)
}

# Plot observations using points and groups as stars with no labels.
# Place tests results in the bottom left corner and give a custom title.
biplot_3d(pca, groups = iris$Species,
          point_type = "point", star_label_alpha = 0,
          show_group_legend = TRUE, group_legend_title = "",
          show_test = TRUE, test_text = resultText(irisTests),
          test_fig = c(0.01,0.5,0.07,0.37),
          show_axes = FALSE, view_theta = 340,
          title = "testing setosa separation")

# Draw a segment between setosa and the other species centroids
x <- c(mean(pca$scores[iris$Species=="setosa", 1]),
       mean(pca$scores[iris$Species!="setosa", 1]))
y <- c(mean(pca$scores[iris$Species=="setosa", 2]),
       mean(pca$scores[iris$Species!="setosa", 2]))
z <- c(mean(pca$scores[iris$Species=="setosa", 3]),
       mean(pca$scores[iris$Species!="setosa", 3]))
lines3d(x, y, z, color = "purple", lwd = 5)

}

}

