% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_device_setup.R
\name{calc_aspect}
\alias{calc_aspect}
\title{Calculate aspect}
\usage{
calc_aspect(x, y, z)
}
\arguments{
\item{x, y, z}{numeric vectors representing point coordinates.}
}
\value{
Returns a vector with three numerical values equal or greater than 1.
}
\description{
Calculate the aspect or the relative scale of data in three dimensions.
}
\examples{
calc_aspect(0:100, 0:50, 0:10)

}

