% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_arrows.R
\name{covArrows3D}
\alias{covArrows3D}
\title{Places a set of arrows in a rgl device}
\usage{
covArrows3D(x, y, z, variable_names = names(x), center_pos = c(0, 0, 0),
  color = "darkorange", head_shape_theta = pi/6, head_shape_n = 3,
  head_size = 0.1, body_length = 1, body_width = 1,
  label_color = "black", label_cex = 1, label_family = "serif",
  label_font = 2, label_adj = 0, label_alpha = 1)
}
\arguments{
\item{x, y, z}{Numeric vectors representing point coordinates where arrows
are pointing to.}

\item{variable_names}{Character vector with the names of the variables
to be plotted as arrow labels.}

\item{center_pos}{Numeric vector of length three containing the 3D
position from which arrows are drawn.}

\item{color}{The color or colors to be used in arrows.}

\item{head_shape_theta}{Numeric, the angle of the barb of the arrows
head ('theta' in \code{\link[heplots]{arrow3d}}).}

\item{head_shape_n}{Numeric, the number of barbs in the arrows head
('n' in \code{\link[heplots]{arrow3d}}).}

\item{head_size}{Numeric, the size of the arrows head ('s' in
\code{\link[heplots]{arrow3d}}).}

\item{body_length}{Numeric, the length of the arrows body, as a fraction
of the true distance from the origin.}

\item{body_width}{Numeric, the width of the arrows body ('lwd' in
\code{\link[heplots]{arrow3d}}).}

\item{label_color}{The color or colors of arrows' labels.}

\item{label_cex, label_family, label_font, label_adj, label_alpha}{The text parameters and the alpha of the arrows' labels
(\code{\link[rgl]{text3d}}, \code{\link[rgl]{rgl.material}}).}
}
\description{
Places a set of arrows in a rgl device
}
\examples{

\dontrun{

# Use iris data
data("iris")

# Principal Components Analysis
pca <- princomp(iris[, 1:4])

# Initializes the rgl device
rgl_init(theta = 330)

# add axes and bounding box
rgl_format(x = pca$scores[, 1], y = pca$scores[, 2], z = pca$scores[, 3],
           axis_titles = c("PCA-1", "PCA-2", "PCA-3"))

# Add data points
points3d(x = pca$scores[iris$Species=="setosa", 1],
         y = pca$scores[iris$Species=="setosa",2],
         z = pca$scores[iris$Species=="setosa", 3],
         color = "green")
points3d(x = pca$scores[iris$Species=="versicolor", 1],
         y = pca$scores[iris$Species=="versicolor",2],
         z = pca$scores[iris$Species=="versicolor", 3],
         color = "red")
points3d(x = pca$scores[iris$Species=="virginica", 1],
         y = pca$scores[iris$Species=="virginica",2],
         z = pca$scores[iris$Species=="virginica", 3],
         color = "blue")

# Add covariance arrows
covArrows3D(x = pca$loadings[,1],
            y = pca$loadings[,2],
            z = pca$loadings[,3],
            body_width = 5)

}

}
\seealso{
\code{\link[heplots]{arrow3d}},\code{\link{biplot_3d}}
}

