% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_group_representation.R
\name{ellipsoid_3d}
\alias{ellipsoid_3d}
\title{Draw an labeled custom ellipsoid}
\usage{
ellipsoid_3d(x, y, z, color = "black", type = "wire and shade",
  level = 0.95, label = "", wire_color = NULL, wire_alpha = 0.2,
  wire_lit = FALSE, shade_color = NULL, shade_alpha = 0.1,
  shade_lit = FALSE, label_color = NULL, label_cex = 2,
  label_family = "sans", label_font = 2, label_adj = c(-0.25, 0.5),
  label_alpha = 1)
}
\arguments{
\item{x, y, z}{Numeric vectors representing point coordinates.}

\item{color}{The default color to be used in all elements.}

\item{type}{Character, representing the type of ellipsoid filling: "wire",
"shade" or "wire and shade" (\code{\link[rgl]{wire3d}},
\code{\link[rgl]{shade3d}}).}

\item{level}{Numeric, the confidence level of a simultaneous confidence
region
(\code{\link[rgl]{ellipse3d}}).}

\item{label}{Character, The ellipsoid label.}

\item{wire_color, shade_color, label_color}{The specific colors to be used
in the ellipsoid's elements. If not NULL, they override the
\code{color} argument.}

\item{wire_alpha, wire_lit}{Numeric, the wire alpha and lit parameters
(\code{\link[rgl]{rgl.material}}).}

\item{shade_alpha, shade_lit}{Numeric, the shade alpha and lit parameters
(\code{\link[rgl]{rgl.material}}).}

\item{label_cex, label_family, label_font, label_adj, label_alpha}{The group
labels text parameters (\code{\link[rgl]{text3d}},
\code{\link[rgl]{rgl.material}}).}
}
\description{
Compute and draw a labeled ellipsoid in a rgl device.
}
\examples{

\dontrun{

# Use iris data
data("iris")

# get setosa
setosa <- iris[iris$Species == "setosa",]

# Initializes the rgl device
rgl_init(zoom = 0.75)

# add axes and bounding box
rgl_format(setosa$Sepal.Length, setosa$Sepal.Width, setosa$Petal.Length,
           axes_titles = c("Sepal length", "Sepal width", "Petal length"))

# Add data points
points3d(setosa[, 1], setosa[, 2], setosa[, 3], color = "black")

# Add ellipsoid
ellipsoid_3d(setosa[, 1], setosa[, 2], setosa[, 3], label = "setosa",
            wire_color = "green", shade_color = "red",
            label_color = "blue", label_adj = c(-1, 0.5))

remove(setosa)

}

}
\seealso{
\code{\link{ellipsoid_3d}}, \code{\link{rgl_init}},
   \code{\link[rgl]{ellipse3d}}, \code{\link[rgl]{wire3d}},
   \code{\link[rgl]{shade3d}}, \code{\link[rgl]{texts3d}}
}
