% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_arrows.R
\name{filter_arrows}
\alias{filter_arrows}
\title{Filter covariance arrows for biplots}
\usage{
filter_arrows(loadings, mim_arrow_dist = 0.5, dimensions = 2)
}
\arguments{
\item{loadings}{Data frame, loadings or the covariances
between variables and the coordinates calculated.}

\item{dimensions}{Numeric, the number of dimensions where arrows will be projected.}

\item{min_arrow_dist}{Numeric, ratio representing the minimun distante
of arrow's points to their origin. 0 return no variables.}
}
\description{
Filter the variables most represented in the firsts dimensions of an ordination object.
}

