% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot3Dim.R
\name{get_colors}
\alias{get_colors}
\title{Get colors for the different levels of a factor variable}
\usage{
get_colors(groups, group_col = palette())
}
\arguments{
\item{groups}{A factor variable containing the group of each observation.}

\item{colors}{The default colors to be used.}
}
\value{
Return a vector of colors matching the levels of groups.
}
\description{
Get colors for the different levels of a factor variable
}

