% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda.R
\name{get_lambda}
\alias{get_lambda}
\title{Calculate lambda (biplot)}
\usage{
get_lambda(sdev, n.obs, dimensions = 2, scale = 1, pc.biplot = FALSE)
}
\arguments{
\item{sdev}{Data frame containing the stardand deviations
along n dimensions}

\item{n.obs}{number of observations (rows) in the original data}

\item{dimensions}{The number of dimensions}

\item{scale}{Numeric, the value defining the
degree in which distances between observations have
priority over distances between variables
(0 = observation-focused, 1 = variable-focused).
It corresponds to the argument \code{scale} in
\code{\link[stats]{biplot.princomp}}.}

\item{pc.biplot}{Character, indicating the type of
biplot transformation of data: "default" and
"pc.biplot", corresponding to the transformations
performed in \code{\link[stats]{biplot.princomp}}
with \code{pc.biplot = FALSE} ("default") and
\code{pc.biplot = TRUE} ("pc.biplot"). If NULL,
no processing is performed, assuming that data
within \code{ordination_object} was previously transformed.}
}
\description{
Calculate the lambda factor to scale biplot dimensions.
The code is a snippet of \code{\link[stats]{biplot.princomp}}).
}
