% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_device_setup.R
\name{rgl_init}
\alias{rgl_init}
\title{Initializes a rgl device}
\usage{
rgl_init(new_device = FALSE, bg_color = "white", view_theta = 15,
  view_phi = 20, view_fov = 60, view_zoom = 0.8, width = 800,
  height = 600)
}
\arguments{
\item{new_device}{If there is a rgl device open, should yet another be opened?}

\item{bg_color}{The background color (\code{\link[rgl]{bg3d}}).}

\item{view_theta, view_phi, view_fov, view_zoom}{The theta and phi angles
(polar coordinates), the field-of-view angle,
and the zoom level of the viewpoint (\code{\link[rgl]{view3d}}).}

\item{width, height}{The width and height of the rgl device window in pixels
(\code{\link[rgl]{par3d}}).}
}
\description{
Initializes a new rgl device using specific settings.
}
\note{
This function is based on the tutorial "A complete guide to 3D
   visualization device system in R - R software and data
   visualization" available in sthda.com Web site and last accessed in may
   24 2016 (\url{http://www.sthda.com/english/wiki/a-complete-guide-to-3d-visualization-device-system-in-r-r-software-and-data-visualization})
}
\examples{

\dontrun{

rgl_init()
rgl_init(bg = "black")
rgl_init(width = 400, height = 300)

}

}
\seealso{
\code{\link[rgl]{bg3d}}, \code{\link[rgl]{view3d}},
   \code{\link[rgl]{par3d}}, \code{\link[rgl]{rgl.open}},
   \code{\link[rgl]{rgl.close}}, \code{\link[rgl]{rgl.clear}}
}
